Add files via upload
This is a development of the previous version of image resizing program. The problem with the previous one was that it can only work on  a specific computer as it does not receive target directory from the user to save the new resized file. No a user can choose the directory in which to save processed images